Rename Data object to DataQlab2
The filename has also been changed.
In this file all different data formats should be stored as classes,
similar to DataQlab2.

Creating a super data class and inherit the specific data formats
from it was the original restructure plan.
However, it seems impossible to define a common set of methods
to extract the data from the file.

Instead a method in eval_ps.py will be defined.
This method checks whether the data object has the needed attributes.