Resolve inheritance from dataclass
Before EvalPowerSeries inherited from the data base class.
(Data from data_import.py)
This has been changed to DataQlab2 from data_formats.py.

Now the DataObj is passed into the init method.
Furthermore, it is checked if the DataObj satisfies the requirements.
(has needed attributes of the correct datatype)
Moreover, the DataObj must be a part of a list of specified data models.
This, and the added setter/getter dataModel should be used
in the future to change the dataModel.

Currently, dataModel can only be set via the config file.
And there is only 1 possible value (Qlab2).

ps_tool.py needs to be adapted accordingly (not done yet).