Adapt init method
As EvalPowerSeries takes a DataObj instead of a filePath now,
the init method of PowerSeriesTool has to be changed accordingly.

First the ini file is read to get the data model that should be used.
Then the DataObj is created and fed into EvalPowerSeries.

When there is need to add more data models,
an input method should be created to change the data model in runtime.
Moreover, the data should be loaded new for these circumstances.
Therefore some raise error statements in eval_ps.py need to be removed.
And also in data_formats.py.
The best option would probably be to use an try/except block.
Then when the data load fails one can change the data model.