Calculate beta for different fixed xi
Reads the outut file beta_fit.csv to get information about xi.
The script calculates beta for the estimated fixed xi.
Here the extracted Q-factor is used.

Moreover, beta is calculated for different xi to sample a distribution.
This is used to quantify the effet of a change in xi.
Therefore, the fitting is performed several times
for equally distributed fixed xi values in the range [xiMin, xiMax].
xiMin and xiMax are read from beta_fit.csv.

What still needs to be done:
- think about the final uncertainty of beta (std?)
- think about xiMin, xiMax values
- use also fixed original xi as fit?